feat(index): Add Introduction Page
[index.md](https://github.com/tylertitsworth/designing-the-dungeon/blob/main/docs/index.md) includes the introduction page to the project. This will be the first page that users see when they visit the site.

fix(.pre-commit-config.yaml): Remove check-added-large-files hook

The `check-added-large-files` hook was removed from the `.pre-commit-config.yaml` file. This hook would prevent larger images and gifs from being uploaded, and based on the scope of this repository that would become necessary.

perf(README.md) Add hyperlink to main readme title

The title of the README.md file is now a hyperlink that will take users to the main page of the project.

feat(mkdocs.yml) Customize theme

Add a logo, font choice, and color scheme to the project